Windows support
Accessing the raw data from the Emotiv dongle on Windows is nearly
identical to how it's done on Mac, thanks to the multi platform support
from hidapi.

The Emotiv dongle has two entries in the list of hid devices, and on
Windows we would consistently try to read data from the wrong one. The
right one can apparently be identified by its interface number, which
should be 1.

[Issue: #1]